Huge commit to add rudimentary support for a query-like API to the ECS.
This puts the repo back on rust nightly, using unboxed closures and the real TraitObject struct as opposed to a hacky reimplementation of it.

The stage is also set for the deferred processing changes to be made, and after that, the tests and examples will be re-written. Hopefully, these changes will not compromise performance too much.